Use new refactored modules from ebabel npm module.
